refactor ChartJs component into a functional component. Add coments to the code
