+++ client/src/components/App.js
@@ -1,55 +1,34 @@
 import React, { Component } from 'react';
 import NavBar from './NavBar';
+import ChartJs from './ChartJs';
+import ReactCharts from './ReactCharts';
+
 import Login from './Login';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import '../css/App.css';
 import { appConfig } from '../utils/constants'
 import { UserSession } from 'blockstack';
 import Button from 'react-bootstrap/Button';
-import { configure, User, getConfig } from 'radiks';
 
-const userSession = new UserSession({
-  appConfig: appConfig
-})
-
-configure({
-  apiServer: 'http://localhost:1260', // TODO this will change to wherever our MongoDB cluster will be hosted in prod
-  userSession
-})
+const userSession = new UserSession({ appConfig: appConfig })
 
 export default class App extends Component {
 
-  constructor(props) {
-    super(props)
-    this.userSession = new UserSession({ appConfig })
-  }
-
-  componentDidMount = async() => {
-    configure({
-      apiServer: 'http://localhost:1260',
-      userSession: this.userSession,
-    });
-    const { userSession } = getConfig();
+  componentDidMount() {
     if (userSession.isSignInPending()) {
-      await userSession.handlePendingSignIn();
-      await User.createWithCurrentUser();
-      window.location = '/';
+      userSession.handlePendingSignIn().then((userData) => {
+        window.history.replaceState({}, document.title, "/")
+        this.setState({ userData: userData })
+      });
     }
   }
 
-  handleSignIn = async(e) => {
-    const { userSession } = getConfig()
+  handleSignIn(e) {
     e.preventDefault()
-    if (userSession.isSignInPending()) {
-      await userSession.handlePendingSignIn();
-      await User.createWithCurrentUser();
-      window.location = '/';
-    }
     userSession.redirectToSignIn()
   }
 
-  handleSignOut = (e) => {
-    const { userSession } = getConfig();
+  handleSignOut(e) {
     e.preventDefault();
     userSession.signUserOut(window.location.origin);
   }
@@ -59,13 +38,9 @@ export default class App extends Component {
       <div className="App">
         {!userSession.isUserSignedIn() ?
           <Login handleSignIn={this.handleSignIn} />
-          :
-          <div>
-            <NavBar />
+          : <div>
             <Button onClick={this.handleSignOut}>Sign Out</Button>
-            <div>
-              <p>Welcome, { userSession.loadUserData().profile.name}</p>
-            </div>
+            <NavBar />
           </div>
         }
       </div>
